LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpTransceiver-setDirection-expected.txt


FAIL setDirection should change transceiver.direction assert_equals: expected (object) null but got (undefined) undefined
PASS setDirection with same direction should have no effect 
FAIL setDirection should change transceiver.direction independent of transceiver.currentDirection assert_equals: expected (object) null but got (undefined) undefined
